Revert "feat(net): use comparingLong method for sortPeers"
This reverts commit c2aea6b81fc23964e214737cb27f8eebeca3d20a.

+++ framework/src/main/java/org/tron/core/net/peer/PeerManager.java
@@ -94,7 +94,7 @@ public class PeerManager {
   }
 
   public static synchronized void sortPeers() {
-    peers.sort(Comparator.comparingLong(c -> c.getChannel().getAvgLatency()));
+    peers.sort(Comparator.comparingDouble(c -> c.getChannel().getAvgLatency()));
   }
 
   public static PeerConnection getPeerConnection(Channel channel) {
